ResourceHandle is now unsigned long long int to prevent overflows in long-running apps Node now uses a NameGenerator to generate names for the same reason

+++ OgreMain/include/OgreResource.h
@@ -37,7 +37,7 @@ Torus Knot Software Ltd.
 
 namespace Ogre {
 
-    typedef unsigned long ResourceHandle;
+    typedef unsigned long long int ResourceHandle;
 
 
 	// Forward declaration
